Fix duplicate import
